fuck yes

+++ code/_core/world/subsystems/cleaning.dm
@@ -4,16 +4,22 @@ var/list/objects_to_delete = list()
 	name = "Cleanup Subsystem"
 	desc = "Cleans up things that need to be deleted."
 	priority = SS_ORDER_DELETE
-	tick_rate = DECISECONDS_TO_TICKS(1)
+	tick_rate = SECONDS_TO_TICKS(1)
 
 /subsystem/delete/on_life()
 
 	for(var/datum/object_to_delete in objects_to_delete)
+
+		if(object_to_delete.qdeleting)
+			objects_to_delete -= object_to_delete
+			continue
+
 		var/time_to_delete = objects_to_delete[object_to_delete]
 		if(time_to_delete > curtime)
 			continue
+
 		objects_to_delete -= object_to_delete
-		world.log << length(objects_to_delete)
+
 		qdel(object_to_delete)
 
 	return TRUE
